+++ protected/humhub/modules/topic/widgets/TopicLabel.php
@@ -10,15 +10,16 @@ namespace humhub\modules\topic\widgets;
 
 
 use humhub\modules\content\models\ContentTag;
+use humhub\modules\topic\models\Topic;
 use humhub\widgets\Label;
 use humhub\widgets\Link;
 use yii\helpers\Html;
 
 class TopicLabel extends Label
 {
-    public static function forTopic(ContentTag $topic)
+    public static function forTopic(Topic $topic)
     {
-        $link = Link::withAction('', 'topic.addTopic')->options(['data-topic-id' => $topic->id]);
+        $link = Link::withAction('', 'topic.addTopic')->options(['data-topic-id' => $topic->id, 'data-topic-url' => $topic->getUrl()]);
         return static::defaultType($topic->name)->sortOrder(20)->color($topic->color)->withLink($link)->icon('fa-star');
     }
 }